# ADD - follow, profile

+++ src/main/java/com/hotkimho/realworldapi/service/FollowService.java
@@ -3,10 +3,12 @@ package com.hotkimho.realworldapi.service;
 
 import com.hotkimho.realworldapi.domain.Follow;
 import com.hotkimho.realworldapi.domain.User;
+import com.hotkimho.realworldapi.dto.follow.FollowResponse;
 import com.hotkimho.realworldapi.dto.user.UserDto;
 import com.hotkimho.realworldapi.exception.DefaultErrorException;
 import com.hotkimho.realworldapi.repository.FollowRepository;
 import com.hotkimho.realworldapi.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
@@ -20,12 +22,13 @@ public class FollowService {
         this.followRepository = followRepository;
     }
 
-    public UserDto followUser(Long followerId, Long followeeId) {
+    @Transactional
+    public FollowResponse followUser(Long followerId, Long followeeId) {
         // 이미 팔로우 했는지 확인
-        if (followRepository.existsByFollowerUserIdAndFolloweeUserId(followerId, followeeId)) {
+        if (followRepository.existsByFollowerIdAndFolloweeId(followerId, followeeId)) {
             User followee = userRepository.findById(followeeId)
                     .orElseThrow(() -> new DefaultErrorException(HttpStatus.NOT_FOUND, "팔로우 대상이 존재하지 않습니다."));
-            return new UserDto(followee, true);
+            return new FollowResponse(followee, true);
         }
 
         User follower = new User(followerId);
@@ -33,7 +36,9 @@ public class FollowService {
 
         try {
             Follow savedFollow = followRepository.save(new Follow(follower, followee));
-            return new UserDto(savedFollow.getFollowee(), true);
+            User followedUser = userRepository.findById(followeeId)
+                    .orElseThrow(() -> new DefaultErrorException(HttpStatus.NOT_FOUND, "팔로우 대상이 존재하지 않습니다."));
+            return new FollowResponse(followedUser, true);
 
         }
         catch (Exception e) {
@@ -42,10 +47,14 @@ public class FollowService {
     }
 
     public void unfollowUser(Long followerId, Long followeeId) {
+        if (!followRepository.existsByFollowerIdAndFolloweeId(followerId, followeeId)) {
+            return;
+        }
+
         try {
             followRepository.deleteByFollowerUserIdAndFolloweeUserId(followerId, followeeId);
         } catch (Exception e) {
-            throw new DefaultErrorException(HttpStatus.UNPROCESSABLE_ENTITY, "팔로우 취소에 실패했습니다.");
+            throw new DefaultErrorException(HttpStatus.UNPROCESSABLE_ENTITY, "팔로우 취소에 실패했습니다. : "+ e.getMessage());
         }
     }
 }
